extract RegexFilter creation to extension method
