+++ main/OpenCover.Framework/OpenCover.Framework.csproj
@@ -85,6 +85,7 @@
       <Link>Properties\Version.cs</Link>
     </Compile>
     <Compile Include="Bootstrapper.cs" />
+    <Compile Include="Filtering\AssemblyAndClassFilter.cs" />
     <Compile Include="Filtering\FilterHelper.cs" />
     <Compile Include="Filtering\FilterType.cs" />
     <Compile Include="Filtering\RegexFilter.cs" />
